Simplify PME spread pipelining stream dependencies
The PME stream needs to wait for the completion of the kernels
submitted to the pipeline streams. This is done via creating
dependencies on those streams, which we do after all kernels are
submitted, so that the API overhead of submitting the kernels does not
delay the launch of subsequent spread kernels in other stream.

In rare corner cases, it is possible that pipelines do not have work
and so do not submit a kernel, and it is most efficient to avoid
creating a useless stream dependency in that case. But in such cases,
the cost of adding and resolving such extra dependencies typically
overlaps with the execution of remaining spread-pipeline kernels. So
instead we keep the GROMACS code simple, by adding a stream dependency
on all pipeline streams regardless of whether a kernel was submitted
to them. This also avoids the minor cost of having book-keeping on the
 hot path to submission of any of the spread-pipeline kernels.

Refs #3988
Refs #5132