Fix comment in decidesimulationworkload.cpp
This comment was wrong when it was written (and reviewed) in 36767396
because mdatoms is a domain-local data structure. A domain can consist
only of non-perturbed atoms and on that domain nPerturbed should be
zero and the haveNonbondedFreeEnergyWork field should be false. In a
subsequent partitioning, the domain can have perturbed atoms thanks to
diffusion, and then the value of this field will change.

The documentation string for the haveNonbondedFreeEnergyWork field is
correct.

Fixes #5484