SYCL: only detect relevant device types
In a89c8a48ae5 (!4990), the behavior was changed to
detect and report all SYCL devices. Since we anyway reject all CPU
devices unless `GMX_SYCL_ALLOW_ALL_DEVICES` is set, we might as well not report
them. Makes the logging behavior  consistent with the SYCL backend in
GROMACS 2025 and with the OpenCL backend.