Make AMBER LEaP atom reordering less verbose
A user should know that their molecular topology changed when between a
top and a tpr file. For that purpose, it is enough to print only the
summary containing the number of reordered dihedrals, number of those
where ordering is kept, and the number of dihedral types.

Furthermore, we limit this only to interaction types where at least one
dihedral was considered for reordering. For most systems, only periodic
improper dihedrals get reordered, which reduces the output to a single
message.

Since the original verbosity is basically a necessity for force field
development and porting (as this information can be hard to extract
from gmx dump), keep it available behind an environment variable.

Refs #4998