Fix sporadic MdrunFEPTest crash/error during CUDA runtime shutdown
Avoid using `TestHardwareEnvironment` since it does not work
well with `mdrun`: it, in CUDA, resets the device context
at the end of the run, invalidating `DeviceStream`s stored in
`TestDevice`s stored in `TestHardwareEnvironment`.

Additionally, when built with CUDA, we were calling `mdrun`
once with (implicit) `-nbfe cpu` and then, for each GPU,
with `-nbfe gpu` (which all were using the same default
GPU allocation), but we were only verifying the output
of the last `mdrun`. Now, we only call `mdrun` once: on GPU
if possible, and on CPU otherwise. Ideally, we should
parameterize the test to try different workload flag
combinations, but not now.

Follow-up to ba9b0deda653f6ece09a07a6523c93abb96ec294.