CUDA: Handle kernels from different architectures slightly better
CUDA kernels are often specialized based on target architecture.

Kernel specialization is done via `GMX_PTX_ARCH` = `__CUDA_ARCH__`.
This, however, means that if a PTX built for one architecture is
ran on another architecture, bad things can happen.

The two notable cases both concern the Prune kernel:
- Block size limits: if the kernel is built for an arch. with
  higher max. threads per SM limit, it will throw an error at
  runtime.
- Shared memory usage: if the kernel is built for an arch. that
  needs shmem for prefetching but is run on a device which is
  not supposed to use prefetching, it will be launched with
  insufficient amount of shared memory and will likely crash.

We cannot directly query whether we use the kernel optimized
for the current device architecture or are JITting a PTX tuned
for a different architecture, so we use an indirect approach.

- Use a macro defined by our CMake to get the list of targets.
- If the current device is not there but can still run kernels,
  print a note to the log file.
- When calculating shmem size for the Prune kernel, err on the
  side of caution when there is a risk.

This approach solves some of the problems that can trip one up
when an older build-tree, configured using pre-63b52fd0ce code,
is re-built for newer GPUs: old CMake cache usually has
`CMAKE_CUDA_ARCHITECTURES=52`, which requires Cj prefetching in
Prune, but the runtime device detection for any modern GPU says
that Cj prefetching is not needed.
Previously, `cudaErrorIllegalAddress` was thrown.
Now, we ask for enough shmem for it to work.

The solution has two downsides:
- We can still crash due to "threads-per-SM" limit (e.g., building
  for sm_52 and running on sm_86).
- We use unnecessarily much shmem (e.g., building for sm_80 and
  running on sm_90).

So, we additionally print a note to the log file recommending the
user to rebuild GROMACS.

Refs: #5444, #5503