Always link executables with SYCL in SYCL build
The resolution of #5393 fixed the known problem with static linking of
GROMACS executables, and specifically did not change the linking of
dynamically-linked GROMACS executables to keep release-2025 behaviour
stable. However, it is simpler to maintain a SYCL build if we always
link with a SYCL-aware tool, and possibly more future-proof, too.

This commit resolves several TODO items from the fix for #5393.

Refs #5393